requirements.txt
click==8.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3

# Flask 核心依赖
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
WTForms==3.0.1
python-dotenv==1.0.0

# 数据库依赖
SQLAlchemy>=1.4.18
alembic>=1.7.0